Refactor T8ditor: split god file, add EditorWorld, extract sub-app panels
Break up the 11.7k-line EditorApp.cpp into cohesive translation units with
no behavior change (each phase build-verified, x64 Release full solution):

- Phase 0: remove dead main_fixed.cpp; sync vcxproj.filters + SceneGraph.h
- Phase 1: extract stateless helpers -> EditorMath.{h,cpp}, EditorUtil.{h,cpp}
- Phase 2: extract scene<->runtime conversions -> EditorSceneSerialization.{h,cpp}
- Phase 3: introduce EditorWorld data model (GetEditorWorld()); route scene data
  (objects, cameras, lights, physics entities, groups, selection, undo) through it
- Phase 4a: extract Play Scene methods -> PlayScenePanel.cpp
- Phase 4b: extract Mesh Editor methods -> MeshEditorPanel.cpp
- Add EditorInternal.h to share editor-private helpers across the split TUs

Sub-app methods were moved byte-for-byte (partial class across TUs). EditorApp.cpp
drops from 11,701 to 10,183 lines. Runtime-verified after phases 3, 4a and 4b.
