+++ T850/T8ditor/CMakeLists.txt
@@ -10,6 +10,12 @@ add_executable(T8ditor
 	EditorMesh.cpp
 	EditorScene.cpp
 	EditorSceneGizmos.cpp
+	EditorMath.cpp
+	EditorUtil.cpp
+	EditorSceneSerialization.cpp
+	EditorWorld.cpp
+	PlayScenePanel.cpp
+	MeshEditorPanel.cpp
 	EditorImGui.cpp)
 
 t850_apply_app(T8ditor)
