+++ T850/T8ditor/EditorWorld.cpp
@@ -1,0 +1,14 @@
+/*********************************************************
+ * T8ditor — EditorWorld instance accessor. See header.
+ *********************************************************/
+
+#include "EditorWorld.h"
+
+namespace t8ditor {
+
+EditorWorld& GetEditorWorld() {
+  static EditorWorld world;
+  return world;
+}
+
+} // namespace t8ditor
